+++ js/main.js
@@ -2,4 +2,4 @@ import './data.js';
 import './modal-shown-events.js';
 import './picture.js';
 import './full-size-image.js';
-// import './full-size-events.js';
+import './fullsize-image-events.js';
